Deduplicate proxy channel creation logic
Inheritance appears to be how shared logic is handled with the
ChannelPoolHanders, so this commit creates a ProxyChanelPoolHandler that the
NoopXXXChannelPoolHanders extend.
