+++ ratpack-core/src/main/java/ratpack/http/client/internal/NoopFixedChannelPoolHandler.java
@@ -17,56 +17,22 @@
 package ratpack.http.client.internal;
 
 import io.netty.channel.Channel;
-import io.netty.channel.pool.AbstractChannelPoolHandler;
-import io.netty.handler.proxy.HttpProxyHandler;
-import io.netty.handler.ssl.SslContext;
-import io.netty.handler.ssl.SslContextBuilder;
 import io.netty.handler.timeout.IdleStateHandler;
-import ratpack.api.Nullable;
 
-import java.net.InetSocketAddress;
-import java.net.SocketAddress;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
 
-public class NoopFixedChannelPoolHandler extends AbstractChannelPoolHandler implements InstrumentedChannelPoolHandler {
+public class NoopFixedChannelPoolHandler extends ProxyChannelPoolHandler implements InstrumentedChannelPoolHandler {
 
   private static final String IDLE_STATE_HANDLER_NAME = "idleState";
 
   private final String host;
   private final Duration idleTimeout;
-  private final ProxyInternal proxy;
-
-  @Nullable
-  private final SslContext sslContext;
 
   public NoopFixedChannelPoolHandler(HttpChannelKey channelKey, Duration idleTimeout) {
+    super(channelKey);
     this.host = channelKey.host;
     this.idleTimeout = idleTimeout;
-    this.proxy = channelKey.proxy;
-    this.sslContext = channelKey.sslContext;
-  }
-
-  @Override
-  public void channelCreated(Channel ch) throws Exception {
-    if (proxy != null && proxy.shouldProxy(host)) {
-      SocketAddress proxyAddress = new InetSocketAddress(proxy.getHost(), proxy.getPort());
-
-      if (proxy.useSsl()) {
-        SslContext sslCtx = sslContext;
-        if (sslContext == null) {
-          sslCtx = SslContextBuilder.forClient().build();
-        }
-        ch.pipeline().addLast(sslCtx.newHandler(ch.alloc(), proxy.getHost(), proxy.getPort()));
-      }
-
-      if (proxy.getCredentials() != null) {
-        // HttpProxyHandler throws a NPE passed a null username or a null password.
-        ch.pipeline().addLast(new HttpProxyHandler(proxyAddress, proxy.getCredentials().getUsername(), proxy.getCredentials().getPassword()));
-      } else {
-        ch.pipeline().addLast(new HttpProxyHandler(proxyAddress));
-      }
-    }
   }
 
   @Override
